chore: implement functions to track what images are available from a given source
As a pre-requisite to changing from a Daemonset based install to a Deployment one,
we need to be able to know what images should be applied to a node when a new one
is added to the cluster.  Previously, this would happen because each node would
get a cache pod which would determine this during initialization, however that
will no longer work with a single pod performing all cache operations.